Removed more code duplication and added looping constructs in
GeneratorViewController.

Working on the duplicated code in the blahViewController.m files, I
found that you cannot use the superclass of PasswordViewController and
NoteViewController as datatypes, and as such I cannot pull up methods,
neither can I push down because that would still require the use of a
supertype to abstract this away.  So I have no clue what to do.  :/

Also there seems to be a conflict in the storyboard file.